Migrate project pages to TS
